# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mentora AI API",
    "description": "API versionada do Mentora AI.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/documents": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "List Tenant Documents",
        "operationId": "list_tenant_documents_api_v1_documents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentResponse"
                  },
                  "type": "array",
                  "title": "Response List Tenant Documents Api V1 Documents Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Upload Document",
        "operationId": "upload_document_api_v1_documents_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/status": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Document Status",
        "operationId": "document_status_api_v1_documents__document_id__status_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/reprocess": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Reprocess Document",
        "operationId": "reprocess_document_api_v1_documents__document_id__reprocess_post",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}": {
      "delete": {
        "tags": [
          "documents"
        ],
        "summary": "Delete Document",
        "operationId": "delete_document_api_v1_documents__document_id__delete",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_document_api_v1_documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "PDF original"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_documents_post"
      },
...
      "DocumentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "knowledge_base_id": {
            "type": "string",
            "format": "uuid",
            "title": "Knowledge Base Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Version Id"
          },
          "ingestion_job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Ingestion Job Id"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "knowledge_base_id",
          "filename",
          "content_hash",
          "size_bytes",
          "status",
          "version_id",
          "ingestion_job_id",
          "created_at"
        ],
        "title": "DocumentResponse"
      },
      "DocumentStatus": {
        "type": "string",
        "enum": [
          "UPLOADED",
          "QUEUED",
          "PROCESSING",
          "INDEXED",
          "FAILED",
          "QUARANTINED",
          "SUPERSEDED"
        ],
        "title": "DocumentStatus"
      },
      "DocumentStatusResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "current_stage": {
            "type": "string",
            "title": "Current Stage"
          },
          "attempt_count": {
            "type": "integer",
            "title": "Attempt Count"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "document_id",
          "status",
          "current_stage",
          "attempt_count",
          "error_message",
          "updated_at"
        ],
        "title": "DocumentStatusResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "QUEUED",
          "PROCESSING",
          "COMPLETED",
          "FAILED",
          "QUARANTINED"
        ],
        "title": "JobStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}